processing/requirements.txt
boto3>=1.34.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
psycopg2-binary>=2.9.0
watchtower>=3.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
